Fix compilation error on macOS/arm (#1156)
Add a cast in the new `statfs` test: `libc::fsfilcnt_t` is not `u64` on
all platforms.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

+++ mountpoint-s3/tests/fuse_tests/statfs_test.rs
@@ -11,8 +11,8 @@ fn statfs_test_static_values(creator_fn: impl TestSessionCreator, prefix: &str) 
     assert_ne!(stats.blocks_free(), 0);
     assert_ne!(stats.blocks_available(), 0);
     // These two are values set by us
-    assert_eq!(stats.files(), u64::MAX / 1024);
-    assert_eq!(stats.files_available(), u64::MAX / 1024);
+    assert_eq!(stats.files() as u64, u64::MAX / 1024);
+    assert_eq!(stats.files_available() as u64, u64::MAX / 1024);
     // These are default values from the Default implementation
     assert_eq!(stats.block_size(), 512);
     assert_eq!(stats.name_max(), 255);
